+++ packages/core/src/common/manifests.ts
@@ -1,0 +1,3 @@
+export type { UserMainTarget } from './manifests/user/main'
+export type { UserDatabaseTarget } from './manifests/user/database'
+export type { Project } from './manifests/project'
