+++ packages/core/src/common/manifests/configuration.ts
@@ -1,4 +1,4 @@
-import type { Configuration as UserConfiguration } from '../../../common/configuration'
+import type { Configuration as UserConfiguration } from '../configuration'
 
 export type Configuration = UserConfiguration & {
   /** The path to the gestalt.config.{js,ts} file */
