+++ packages/core/src/common/manifests/plugin.ts
@@ -1,4 +1,4 @@
-import { Plugin as UserPlugin } from '../../../node/plugin'
+import { Plugin as UserPlugin } from '../../node/plugin'
 
 export type Plugin = UserPlugin & {
   name: string
