+++ packages/core/src/common/manifests/project.ts
