+++ packages/core/src/common/manifests/targets.ts
