+++ packages/core/src/common/manifests/targets/main/route.ts
