+++ packages/core/src/common/manifests/user/database.ts
