+++ packages/core/src/common/manifests/user/main.ts
