--- packages/core/src/common/target.ts
@@ -1,2 +1,0 @@
-export type { UserMainTarget } from './targets/main'
-export type { UserDatabaseTarget } from './targets/database'
