+++ packages/core/src/node/project.ts
@@ -1,2 +1,1 @@
 export { loadProject } from './project/load/project'
-export * from './project/models'
