+++ packages/core/src/node/project/load/config.ts
@@ -1,5 +1,5 @@
 import { configurationFileName } from '../../../common/constants'
-import { Configuration } from '../models/configuration'
+import { Configuration } from '../../../common/manifests/configuration'
 import { ModuleLoader } from './module-loader'
 import { findPathUp } from '../../../node/fs'
