+++ packages/core/src/node/project/load/project.ts
@@ -1,4 +1,4 @@
-import { Project } from '../models/project'
+import { Project } from '../../../common/manifests/project'
 import { Abort } from '../../../common/error'
 import { parentDirectory, joinPath } from '../../../node/path'
 import { loadTargets } from './target'
