+++ packages/core/src/node/project/load/target.ts
@@ -1,6 +1,6 @@
 import { glob, joinPath } from '../../../node/path'
 import { mainTargetFileName } from '../../../common/constants'
-import { Targets } from '../models/targets'
+import { Targets } from '../../../common/manifests/targets'
 import { loadMainTarget } from './targets/main'
 import { ModuleLoader } from './module-loader'
