+++ packages/core/src/node/project/models.ts
@@ -1,5 +1,5 @@
-export type { Plugin } from './models/plugin'
-export type { Project } from './models/project'
-export type { Configuration } from './models/configuration'
-export type { MainTarget } from './models/targets/main'
-export type { Targets } from './models/targets'
+export type { Plugin } from '../../common/manifests/plugin'
+export type { Project } from '../../common/manifests/project'
+export type { Configuration } from '../../common/manifests/configuration'
+export type { MainTarget } from '../../common/manifests/targets/main'
+export type { Targets } from '../../common/manifests/targets'
